source/ingestion/requirements.txt
pika==1.3.2
requests==2.31.0
websocket-client==1.7.0